Alternative design for integration backends.
