+++ spacy_llm/backends/__init__.py
@@ -1,6 +1,6 @@
-from .dolly import backend_dolly_hf
-from .langchain import backend_langchain, query_langchain
-from .minichain import backend_minichain, query_minichain
+from .integration import backend_dolly_hf
+from .integration import backend_langchain, query_langchain
+from .integration import backend_minichain, query_minichain
 from .rest import backend_rest
 
 __all__ = [
