+++ spacy_llm/backends/integration/__init__.py
@@ -1,0 +1,13 @@
+from .base import Backend
+from .langchain import backend_langchain, query_langchain
+from .minichain import backend_minichain, query_minichain
+from .dolly import backend_dolly_hf
+
+__all__ = [
+    "Backend",
+    "query_minichain",
+    "query_langchain",
+    "backend_minichain",
+    "backend_langchain",
+    "backend_dolly_hf",
+]
