+++ spacy_llm/backends/rest/__init__.py
@@ -1,3 +1,4 @@
 from .registry import backend_rest
+from .backend.base import Backend
 
-__all__ = ["backend_rest"]
+__all__ = ["backend_rest", "Backend"]
