+++ usage_examples/ner_dolly/dolly_ner_fewshot.cfg
@@ -18,7 +18,7 @@ labels = PERSON,ORGANISATION,LOCATION
 
 [components.llm.task.examples]
 @misc = "spacy.FewShotReader.v1"
-path = "ner_examples.yml"
+path = "examples.yml"
 
 [components.llm.task.normalizer]
 @misc = "spacy.LowercaseNormalizer.v1"
